Cleanup code
Notable Changes:
- fix data leaks
- fix logic errors
- keep on going if you don't get the RUDP FINISH message after you have tried a few times bc it could still succeed anyway

+++ lib/include/chiaki/gkcrypt.h
@@ -29,8 +29,8 @@ typedef struct chiaki_gkcrypt_t {
 	uint8_t index;
 
 	uint8_t *key_buf; // circular buffer of the ctr mode key stream
-	size_t key_buf_size;
-	size_t key_buf_populated; // size of key_buf that is already populated (on startup)
+	uint64_t key_buf_size;
+	uint64_t key_buf_populated; // size of key_buf that is already populated (on startup)
 	uint64_t key_buf_key_pos_min; // minimal key pos currently in key_buf
 	size_t key_buf_start_offset; // offset in key_buf of the minimal key pos
 	uint64_t last_key_pos;        // last key pos that has been requested
